P: Map.cs A: Added a map class to keep track of the map and the lost markers on it.
P: Program.cs
A: Put in code to check for lost markers before making a ship lost
A: Moved CanMoveForward to the Map class.
R: That way the function doesn't have to be static and it doesn't require the xMax and yMax values because the Map class already holds them.

P: UnitTest1.cs
A: Added a test for setting markers on the map.